Fixes server bugs, work on cartographer
All bugs on the server (because of all my previous intialization/algorithms) have been fixed.

The cartographer now receives a list of all units surrounding him, based on instrument level.

Still need to test if it displays and works correctly, but for that I first need to create monsters and position them across the map.