Small fixes, better UI
Small bug fixes, better upgrades (and a way to display what they mean)

Most importantly, work has started on improving the UI. The first mate now shows a red ring to indicate that certain angles are forbidden.

Soon, the sailor will be updated with a completely new screen.

And, once you've upgraded, there will be a piece of text that says "upgrade requested; it takes a turn before you can use it"