Sailor interface (almost) finished
The sailor interface now has all the correct sliders, icons, buttons, and visual niceties (such as showing when something is a forbidden move).

All I need to do is mirror the system on the server. Currently, boats have no speed (and just move based on current sail level).

I also need to determine a good movement range (when it comes to upgrading). Right now, it's too restricted if you can only go -1 or +1 on a turn.