Ocean detection
Taking a break from the backstory simulator (because I need to rewrite/rethink lots of things, and I don't feel like doing that).

Now, oceans are detected on the map, and subsequently saved and named. This will make the map more interesting (as there are actual oceans and names attached), but also help with giving clues about where something is. ("The treasure lies in the Blabla Ocean".)